+++ packages/context-selector/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.8](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-context-selector@0.1.7...@equinor/fusion-react-context-selector@0.1.8) (2022-12-02)
+
+**Note:** Version bump only for package @equinor/fusion-react-context-selector
+
+
+
+
+
 ## [0.1.7](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-context-selector@0.1.6...@equinor/fusion-react-context-selector@0.1.7) (2022-11-29)