+++ storybook/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.4](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@2.0.3...@equinor/fusion-react-components-stories@2.0.4) (2022-12-02)
+
+**Note:** Version bump only for package @equinor/fusion-react-components-stories
+
+
+
+
+
 ## [2.0.3](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@2.0.2...@equinor/fusion-react-components-stories@2.0.3) (2022-11-29)
 
 **Note:** Version bump only for package @equinor/fusion-react-components-stories